Fix profile view when not authed

+++ resources/views/users/profiles/show.blade.php
@@ -42,7 +42,7 @@
                                                             </p>
                                                         </li>
                                                         <li class="col-sm-4">
-                                                            @if (! auth()->user()->isInHorseTeam($horse))
+                                                            @if (auth()->check() && ! auth()->user()->isInHorseTeam($horse))
                                                                 @if (Auth::user()->isFollowing($horse))
                                                                     {{ Form::open(['route' => ['follows.destroy', $horse->id()], 'method' => 'DELETE']) }}
                                                                     <button type="submit" class="btn btn-mint">{{ trans('copy.a.unfollow') . $horse->name() }}</button>
@@ -94,7 +94,7 @@
                                                             </p>
                                                         </li>
                                                         <li class="col-sm-4">
-                                                            @if (! auth()->user()->isInHorseTeam($follow))
+                                                            @if (auth()->check() && ! auth()->user()->isInHorseTeam($follow))
                                                                 @if (auth()->user()->isFollowing($follow))
                                                                     {{ Form::open(['route' => ['follows.destroy', $follow->id()], 'method' => 'DELETE']) }}
                                                                     <button type="submit" class="btn btn-mint">{{ trans('copy.a.unfollow') . $follow->name() }}</button>
@@ -137,19 +137,19 @@
                         <h4 class="text-lg text-overflow mar-no">{{ $user->fullName() }}</h4>
 
                         <div class="pad-ver btn-group">
-                            @if (auth()->user()->facebook())
-                                <a title="" href="http://www.facebook.com/{{ auth()->user()->facebook() }}" target="_blank" class="btn btn-icon btn-hover-primary fa fa-facebook icon-lg add-tooltip" data-original-title="Facebook" data-container="body"></a>
+                            @if ($user->facebook())
+                                <a title="" href="http://www.facebook.com/{{ $user->facebook() }}" target="_blank" class="btn btn-icon btn-hover-primary fa fa-facebook icon-lg add-tooltip" data-original-title="Facebook" data-container="body"></a>
                             @endif
 
-                            @if (auth()->user()->twitter())
-                                <a title="" href="http://www.twitter.com/{{ auth()->user()->twitter() }}" target="_blank" class="btn btn-icon btn-hover-info fa fa-twitter icon-lg add-tooltip" data-original-title="Twitter" data-container="body"></a>
+                            @if ($user->twitter())
+                                <a title="" href="http://www.twitter.com/{{ $user->twitter() }}" target="_blank" class="btn btn-icon btn-hover-info fa fa-twitter icon-lg add-tooltip" data-original-title="Twitter" data-container="body"></a>
                             @endif
 
-                            @if (auth()->user()->website())
-                                <a title="" href="{{ auth()->user()->website() }}" target="_blank" class="btn btn-icon btn-hover-warning fa fa-laptop icon-lg add-tooltip" data-original-title="Website" data-container="body"></a>
+                            @if ($user->website())
+                                <a title="" href="{{ $user->website() }}" target="_blank" class="btn btn-icon btn-hover-warning fa fa-laptop icon-lg add-tooltip" data-original-title="Website" data-container="body"></a>
                             @endif
                         </div>
-                        @if (auth()->user()->id() == $user->id())
+                        @if (auth()->check() && (auth()->user()->id() == $user->id()))
                             <a href="{{ route('users.profiles.edit') }}" class="btn btn-block btn-mint">{{ trans('copy.a.edit') }}</a>
                         @else
                             <a href="{{ route('conversation.create', ['contact' => $user->id()]) }}" class="btn btn-block btn-mint">
